refactor: renaming the files to follow the standard structure
