+++ packages/bruno-app/src/components/Environments/Common/ImportEnvironmentModal/hooks/useEnvironmentTarget/index.js
@@ -1,7 +1,7 @@
 import { useDispatch, useSelector } from 'react-redux';
 import { importEnvironment, saveEnvironment, updateEnvironmentColor } from 'providers/ReduxStore/slices/collections/actions';
 import { addGlobalEnvironment, saveGlobalEnvironment, updateGlobalEnvironmentColor } from 'providers/ReduxStore/slices/global-environments';
-import { normalizeEnvName } from '../utils';
+import { normalizeEnvName } from '../../utils';
 
 export const useEnvironmentTarget = (type, collection) => {
   const dispatch = useDispatch();
